safari-600.7-branch/LayoutTests/canvas/philip/tests/2d.drawImage.incomplete.html

<!DOCTYPE html>
<title>Canvas test: 2d.drawImage.incomplete</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#0f0';
ctx.fillRect(0, 0, 100, 50);
var img = new Image();
img.src = "NotARealFileName.png";
_assertSame(img.complete, false, "img.complete", "false");
ctx.drawImage(img, 0, 0);


});
</script>
