LayoutTests/platform/mac-snowleopard/canvas/philip/tests/2d.drawImage.incomplete-expected.txt

< [index] >
2d.drawImage.incomplete
References: 2d.drawImage.incomplete.image
Actual output:
Expected output:

Failed assertion img.complete === false (got true[boolean], expected false[boolean])
